refactor(nx): do not ask for directory when creating apps and libs
