+++ packages/workspace/src/schematics/library/schema.json
@@ -15,8 +15,7 @@
     },
     "directory": {
       "type": "string",
-      "description": "A directory where the app is placed",
-      "x-prompt": "In which directory should the library be generated?"
+      "description": "A directory where the app is placed"
     },
     "linter": {
       "description": "The tool to use for running lint checks.",
